Graph - introduce abstractions to mediate with simulation
There are some details the simulation needs to know about nodes and
edges. For instance, it needs the edge distance and the node radius.
Before, we were just assuming nodes and edges were just objects with
those properties, copied from props. Now, we at least have an interface
to document what's required to pass to the simulation.