Simulations - refactor interaction between simulation and geometry
Before, there was some mixing of concerns between the simulated entities
and the geometry they took on. You could tell because a
`SimulatedCircle` was required to have an `id`, which is a little
awkward. We want to standardize the communication between the components
and the simulation; they do that by creating simulated objects with
shapes. The logic that is specific to geometry and shapes has been
extracted out to a submodule.