Graphs - replace manual simulation control with component
This commit introduces a component, `SimulatedLayout`, and refactors the
`Graph` component to use that. `SimulatedLayout` handles managing a
simulation for child components, but it gets it's config from those
child components. Each direct child is expected to supply a
`getSimulationConfig` to define the elements, forces, and constraints it
wants to simulate.

This will make it easier for the `Node` component to internally manage
more elements, like a node and a label.