Simulation - make getSimulationConfig support promises
This fixes getting the correct shape for labels. We were having a
problem with labels because the ref wasn't present by the time the
simulated layout component's `componentDidMount` got called and configs
were aggregated. Instead, we make the `getSimulationConfig` return a
promise, and the `SimulatedLayout` waits for the promises to resolve
before creating the simulation.