Simulation Rearchitecture - use context instead of refs
This commit changes the architecture of how `SimulatedLayout` gets its
config from child components. Before, we exposed a `getSimulationConfig`
method on the child components and used refs to call it. That's
problematic for a number of reasons.

First, it wasn't clear when the methods were getting called, and there
were bugs around them getting called at the wrong time. Second, some
components want to define their config based on actual values from
rendering to the dom - for instance, a label will define rectangular
shape in the simulation based on the size of the text.

Lastly and most importantly, the ref based solution made it hard to
write wrapping components. It should be easy to write new components
that render graph elements. With refs, each component that wanted to
wrap graph elements had to forward refs and aggregate configs.

Now, you can just render a `<Node>` and it will register the appropriate
data with the simulation.