Restructure Simulation to use ForceApplications
This commit restructures the way the simulation takes in rules from the
clients. Now the simulation doesn't have to know about all the different
types of rules - a rule is now just a function that returns a set of
`ForceApplication`s to the next state of the simulation. This makes it
dramatically easier and simpler to add in extra positioning rules to a
simulation.