Orienting rule - rework strategy for orienting
This commit changes the strategy for the orienting rule to push elements
towards a point along the orientation's direction the same distance away
as the element. This is better than the old strategy, which just pushed
elements in the orientation's direction, because sometimes elements need
to go up or down to get in the correct orientation.