Label - extract utility for monitoring shape, respond to changes
This commit changes the strategy used by the `Label` element for
registering a shape with the context. Previously, it would pass a ref
callback and post render get the svg element shape to register. Now it
polls on an interval, and updates the shape in the simulation on change.
That allows these elements to rendered in a hidden view first, and then
expanded, as is done by .... whatever presentation framework I use.