PositioningRules - make adjustments for smoother application
This commit changes the orienting and postioning rules to remove the
"jitters". The positioning rule now scales its strength with the square
of the distance to the desired point, scaled to be between 0 and 1. The
orientation rule smoothly applies itself across the range instead of not
applying itself while inside the range.